Move lovelace data to /src/data/lovelace.ts (#2119)
* move lovelace data

* move types

* change imports from cards
