+++ src/panels/lovelace/cards/hui-conditional-card.ts
@@ -1,7 +1,8 @@
 import createCardElement from "../common/create-card-element";
 import { computeCardSize } from "../common/compute-card-size";
 import { HomeAssistant } from "../../../types";
-import { LovelaceCard, LovelaceCardConfig } from "../types";
+import { LovelaceCard } from "../types";
+import { LovelaceCardConfig } from "../../../data/lovelace";
 
 interface Condition {
   entity: string;
