+++ src/panels/lovelace/cards/hui-glance-card.ts
@@ -10,7 +10,8 @@ import { classMap } from "lit-html/directives/classMap";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { hassLocalizeLitMixin } from "../../../mixins/lit-localize-mixin";
 import { HomeAssistant } from "../../../types";
-import { LovelaceCard, LovelaceCardConfig, LovelaceCardEditor } from "../types";
+import { LovelaceCard, LovelaceCardEditor } from "../types";
+import { LovelaceCardConfig } from "../../../data/lovelace";
 import { longPress } from "../common/directives/long-press-directive";
 import { EntityConfig } from "../entity-rows/types";
 import { toggleEntity } from "../common/entity/toggle-entity";
