+++ src/panels/lovelace/cards/hui-picture-glance-card.ts
@@ -5,7 +5,8 @@ import { TemplateResult } from "lit-html";
 import { hassLocalizeLitMixin } from "../../../mixins/lit-localize-mixin";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { DOMAINS_TOGGLE } from "../../../common/const";
-import { LovelaceCard, LovelaceCardConfig } from "../types";
+import { LovelaceCard } from "../types";
+import { LovelaceCardConfig } from "../../../data/lovelace";
 import { EntityConfig } from "../entity-rows/types";
 import { navigate } from "../../../common/navigate";
 import { HomeAssistant } from "../../../types";
