+++ src/panels/lovelace/editor/hui-dialog-save-config.ts
@@ -9,9 +9,12 @@ import { PaperDialogElement } from "@polymer/paper-dialog/paper-dialog";
 import "@polymer/paper-button/paper-button";
 
 import { HomeAssistant } from "../../../types";
-import { LovelaceConfig } from "../types";
 
-import { saveConfig, migrateConfig } from "../common/data";
+import {
+  saveConfig,
+  migrateConfig,
+  LovelaceConfig,
+} from "../../../data/lovelace";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { hassLocalizeLitMixin } from "../../../mixins/lit-localize-mixin";
