+++ src/panels/lovelace/editor/hui-migrate-config.ts
@@ -11,7 +11,7 @@ import { fireEvent } from "../../../common/dom/fire_event";
 
 import { HomeAssistant } from "../../../types";
 import { hassLocalizeLitMixin } from "../../../mixins/lit-localize-mixin";
-import { migrateConfig } from "../common/data";
+import { migrateConfig } from "../../../data/lovelace";
 
 export class HuiMigrateConfig extends hassLocalizeLitMixin(LitElement) {
   protected hass?: HomeAssistant;
