+++ src/panels/lovelace/types.ts
@@ -1,24 +1,5 @@
 import { HomeAssistant } from "../../types";
-
-export interface LovelaceCardConfig {
-  id?: string;
-  type: string;
-  [key: string]: any;
-}
-
-export interface LovelaceViewConfig {
-  title?: string;
-  badges?: string[];
-  cards?: LovelaceCardConfig[];
-  id?: string;
-  icon?: string;
-}
-
-export interface LovelaceConfig {
-  _frontendAuto: boolean;
-  title?: string;
-  views: LovelaceViewConfig[];
-}
+import { LovelaceCardConfig } from "../../data/lovelace";
 
 export interface LovelaceCard extends HTMLElement {
   hass?: HomeAssistant;
